requirements.txt
streamlit>=1.33
pandas>=2.0
numpy>=1.24
plotly>=5.20
kaleido>=0.2.1
seaborn>=0.13
matplotlib>=3.8
fpdf2>=2.7
openpyxl>=3.1